feat: Enhance course management and payment processing
- Updated courseController to include specific folder for thumbnail uploads and improved response statuses.
- Refactored paymentController to utilize userId from request context and updated success/cancel URLs to use environment variable.
- Modified quizController and reviewController to use consistent naming for course model and improved quiz generation logic.
- Enhanced userController to return consistent response statuses and added photo upload handling.
- Integrated helmet for security enhancements in index.js and adjusted CORS settings.
- Improved error handling in isAuth middleware and updated multer configuration for unique file naming.
- Added required field for certificateId in certificate model and changed Price type in course model to Number.
- Removed deprecated questions model and created a new questionsModel for better structure.
- Updated routes to ensure proper authentication and streamlined endpoint naming.
- Added .gitkeep to public directory for maintaining structure.

+++ server/config/cloudnary.js
@@ -1,25 +1,29 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-const uploadOnCloudinary = async (filePath) => {
-  cloudinary.config({
+cloudinary.config({
     cloud_name: process.env.CLOUDINARY_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
     api_secret: process.env.CLOUDINARY_API_SECRET,
     secure: true,
   });
 
+const uploadOnCloudinary = async (filePath , folderName) => {
   try {
     if (!filePath) {
       return null;
     }
     const uploadResult = await cloudinary.uploader.upload(filePath, {
-      resource_type: "auto",
+      folder: `DevTrack/${folderName}`,
+      resource_type: "auto" ,
     });
+
     fs.unlinkSync(filePath);
     return uploadResult.secure_url;
   } catch (error) {
-    fs.unlinkSync(filePath);
+    if (filePath && fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
     console.error("Error uploading image to Cloudinary:", error);
     return null;
   }
